Fix some linter errors
