+++ src/modules/deleteTodoListItems.js
@@ -1,0 +1,9 @@
+function deleteAnItem(i, todos) {
+  const temp = [...todos];
+  todos.splice(i, 1);
+  if (todos === temp) {
+    throw new Error('Error deleting item');
+  }
+  return todos;
+}
+module.exports = deleteAnItem;
