+++ src/modules/index.js
@@ -1,2 +1,2 @@
 import '../style.css';
-import './todo.js';
+import './todoList.js';
